watchers/pyproject.toml
[project]
name = "watchers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.189.0",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.4",
    "playwright>=1.58.0",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
]
